Match `PATH` case insensitively in Windows installer
The Windows installer expands the path to the `C:\Program Files`
directory using the standard casing for this.  However, our `PATH` value
does not need to contain the same case, since on Windows this directory
is case insensitive, and as a result, if the two differ in case, we
don't correctly detect that the Git executable is in `PATH`.

Let's fix this by lowercasing both versions of the text.  Note that
instead of using the `Lowercase` function, which deals only with ASCII
characters, we use `AnsiLowercase`, which, despite its name, uses the
current Windows locale.

This is not guaranteed to produce correct results in all cases because
the file system's case folding is done in the kernel in a
locale-insensitive way based on the tables created when the file system
was created, but there's no way to do better than this in the general
case since we don't have functionality to implement the kernel's case
mapping and it is impossible to correctly fold arbitrary Unicode text
in a locale-insensitive way.  We therefore adopt the least bad option
available while continuing to support Windows.