Implement support for authtype in test credential helper
In newer versions of Git, the credential helper protocol has learned to
support capabilities and credentials that aren't based around a username
and password.  This allows the use of Bearer credentials, and it allows
credential helpers to implement complex functionality like Digest and
NTLM (both of which require insecure password storage) so we don't have
to.

We'd like to support this in our testsuite, so extend our credential
helper to accept these new capabilities and use them to determine a
response.  In conformance with the credential helper protocol, we only
advertise support for capabilities which we ourselves support.

In our files in `CREDSDIR`, instead of looking for two parts separated
by colons which are the username and password, accept three parts, which
are the authentication type, username, and credential.  If the
authentication type is empty, then the credential is the password;
otherwise, the username is empty, and the credential is sent in the
credential field.

Right now, we don't make use of this new functionality, but we do update
existing credential files to make them compatible.  Use of the new
functionality will come in a future commit.