checkout: gracefully handle files deleted from the index
Right now, when someone deletes a pointer from the index with `git rm`
and then runs `git lfs checkout`, the operation fails with a message of
"Could not update the index" because our invocation of `git
update-index` is missing the `--add` flag.

Obviously, the user does not expect an error in this case, and `git
checkout` simply ignores files staged for deletation, so let's do the
same thing.  If a file on disk is deleted, check the index with `git
diff-index` to see if it's deleted from `HEAD`.  If so, ignore the file,
just like Git does. Note that we use `git diff-index` specifically
because it doesn't refresh the index and is therefore much cheaper than
alternatives, such as `git status`, which might do that.