t-clone: add `.gitattributes` before pushing
In two of our test repositories in this file, we never commit the
`.gitattributes` file and instead rely on the fact that `git lfs clone`
sometimes processes all files in the working tree to determine which are
pointer files.  This is inefficient and we don't want to rely on this
behaviour, especially since it differs from that of `git clone`, so fix
our tests so that we explicitly commit the `.gitattributes` file.