Optimize access for scanning trees
Right now, if the user is using partial clone, our call to `git ls-tree`
against HEAD is expensive because `git ls-tree` needs to download each
blob, which it does incrementally instead of all at once.

If we're scanning the tree from HEAD, then we can avoid the expense of
doing this by running `git ls-files` with a pattern that matches only
LFS files, which makes the operation much cheaper, since we avoid
needing to download blobs for many of those objects.  We can format the
data such that it matches the pattern we expect for `git ls-tree` so
that we can avoid modifying most of the calls and continue to let things
function in the same way.  Do so, but limit our changes to Git 2.42.0
and newer, since the `objecttype` argument is new in that version.