subprocess: add a version of BufferedExec with only stdout
In some cases, we call `BufferedExec` and only want standard output;
that is, we ignore standard error.  In such a case, we can cause the
data to be buffered forever and we'll block if there's enough of it,
especially on Windows.

Let's add a `StdoutBufferedExec` that redirects standard error to
`/dev/null`, since that's effectively what we're doing already, except
that our new function will be more robust because it won't cause
indefinite blocking.