lfs: don't wait until the end to process standard error
When we process log output, we can sometimes get a large amount of
data on standard error, which is buffered.  If we wait until the end to
process it, then the child process can block because it's trying to
write to standard error but we're not reading from it, and in turn we
can block because we're waiting for more standard output.

The solution is simple: process standard error in another goroutine and
send the result over a channel to be read at the end.  This guarantees
that we'll always process the data as expected and prevent deadlock.