creds: handle authtype and credential
Now that our test credential helper and test server are capable of
supporting the new authtype and credential fields, let's pass the
appropriate capability to the helper.  We use the fields if they exist,
and pass them back to the `approve` call to the helper on success, along
with the capability.

If these fields don't both exist, we fall back to the username and
password fields.  This allows us to do thing such as query different
accounts with the username if one is specified in the URL, but get the
credentials back in a format that contains a non-Basic response.

Since we now support more than just Basic and Negotiate, let's make sure
we iterate through all of the supported authentication schemes, not just
those two.  This allows the credential helper to support multiple
schemes.