docs/api: note charset parameter may be sent
Since at least commit 099adaaa5e59be83ee9f3079398c4398fd282e50 in 2013
the Git LFS client has included a "charset=utf-8" parameter in some
or all of the Content-Type headers it sends to server API endpoints.

As noted in #5775, we do not at present mention this parameter in
our API documentation, and the "charset" parameter is not required
by any IETF RFCs for MIME types other than "text/*" types, so we
note that servers should be prepared to accept this header parameter
but that it is not required.

h/t KalleOlaviNiemitalo for the suggestion