git-credential-lfstest: create and use a credential struct
We're going to start processing more complex sets of credentials in the
future, so to keep our code nice and tidy, move the credential handling
into a struct so that we can simplify things.  This allows us to return
only two values from `credsForHostAndPath` instead of the current four,
a number which would only increase in the future.