git-credential-lfstest: add basic support for state
The new functionality in Git allows persisting state across requests and
providing multistage responses.  To allow testing that, let's implement
the `state[]` and `continue` arguments.

The `state[]` entry contains one or more items of state that are
prefixed by the credential helper to distinguish them from those from
other credential helpers.  These are passed back to the credential
helper on success, failure, or a second round.

The `continue` boolean value indicates whether, upon receiving a 401
response, authentication should continue for another round.  This flag
allows three-legged auth such as NTLM or Kerberos (via either NTLM or
Negotiate schemes) to be successfully implemented by the credential
helper.

Now that we're writing a serialized credential directly into a variable,
let's make sure we copy the relevant pieces from the original credential
structure, including host, path, protocol, and capabilities, so that
these are passed along.  Except for the capabilities, none of these are
strictly needed since Git will fill them in, but it's still nicer to
provide them in the output for completeness.

We document the possible values of a line in the configuration file to
make them easier to understand as well.