Implement multistage authentication
In 2.46, Git will support multistage authentication from credential
helpers.  This will allow Git and Git LFS to implement functionality
like NTLM and Kerberos in the credential helper, which lets this
functionality to work even if Git LFS doesn't support it natively.

This requires two separate pieces of data.  First, it involves a
`state[]` field, which each credential helper can add to keep track of
state.  Second, it allows the usage of a boolean `continue` field, which
indicates that the response is multistage and this is not the final
stage.

In order to make this work, we adjust a few things.  First, we advertise
the `state` capability.  Additionally, we save and pass back the
`state[]` fields that the credential helper may send to us.  We also
don't change the authentication scheme if the helper told us that this
was a multistage response.  Finally, we add a check to avoid a
credential helper getting stuck in an infinite loop if it keeps handing
back the same credentials.