t: add tests with missing Retry-After header
A prior commit in this PR resolves a bug where a 429 response to an
upload or download request causes a Go panic in the client if the
response lacks a Retry-After header.

We can simulate this condition and confirm that the changes fix the
problem by updating our test server to respond with a 429 but no
Retry-After header when passed a new sentinel value, and then adding
a pair of tests which send this value.  Both tests generate the
Go panic condition without the other changes in this PR.