.github/workflows,script: use custom Git on macOS
Possibly since we first switched to the use of GitHub Actions for our
CI jobs, in PR #3808, the version of Git we've used on macOS in our
build-earliest and build-latest jobs has actually been the one installed
by Homebrew rather than the one we've custom-built for the purpose.

This is due to the fact that, by default, the PATH environment variable
supplied in GitHub Actions runners on macOS has the /opt/homebrew/bin
location listed ahead of the /usr/local/bin location where we install
our custom version of Git.

We can address this by ensuring that our build-git script installs
the custom version of Git into a unique location, when running on
macOS, and then placing this location ahead of all others in the PATH
environment variable.

On Linux we can just continue to install our version of Git into
/usr/local/bin, as that takes precedence over any other installed
version of Git on the GitHub Actions runners.