ssh: fix connection creation "leaking" connections
Commit 44b8801cbb57891b4c66fc3c43551802e5532bd1 introduced
a bug where connections were being created on-demand
but were not being stored in the connections slice.

When trying to end connections at the end of a transfer,
there were no handles to them. As a result, all but the
primary connection of the session were hung with no "quit"
message. LFS clients were simply hung until idle timeout
kicked in.