t: add lfstest-genrandom program for tests
We use several techniques in our test suite to acquire random data
and filenames, of which the most prevalent one is to read from the
/dev/urandom special file.  Recently, this appears to no longer
work reliably when our CI jobs are run on Windows by our GitHub Actions
workflows, possibly due to an upstream change in Git Bash, MSYS2,
Cygwin, or for some other reason.

We also use the uuidgen(1) command in one test in our t/t-prune.sh
test script to populate Git LFS object content.  However, that command
is not installed with Git Bash, so on Windows this data is simply empty;
while the test still passes, this is less than ideal.

We already have a number of helper programs written in Go that provide
platform-agnostic utilities to our test suite, and Go's crypto/rand
package can be used to generate random data on all the platforms
we support.  On Windows in particular it uses the ProcessPrng API,
since /dev/urandom does not exist on that platform.  (We are only able
to make use of /dev/urandom on Windows because a simulation of it is
provided by the MSYS2 environment, and Git Bash, which we use to run
our test suite, is built on MSYS2.)

Therefore we add a new lfstest-genrandom program for our test suite
which uses the crypto/rand Go package to return strings of random data.
If a size is provided as a command-line argument, it returns that
number of random bytes; otherwise, it returns an effectively endless
stream of data.  This parallels the design of the Git test suite's
test-genrandom helper:

  https://github.com/git/git/blob/a7dae3bdc8b516d36f630b12bb01e853a667e0d9/t/helper/test-genrandom.c

When the --base64 command-line flag is specified, our new program returns
data encoded using the regular Base64 specification from RFC 4648.  When
the --base64url flag is specified instead, the returned data is encoded
with the alternate Base64 specification from RFC 4648 that is suitable
for use in filenames and URLs.  In this case, slash and plus characters
are avoided in favour of underscore and dash characters.  With either
flag, the encoded data has no Base64 padding, and if a size was provided
on the command line, the data will be truncated to exactly that number
of bytes after encoding (rather than being truncated before encoding,
which would produce extra output bytes).

While we could choose to only produce the URL-compatible form of Base64
data, we would have to rewrite a number of tests which expect to be
able to read the random data we've written into a test file and then
pass it to our calc_oid() shell function, which calls the Bash builtin
command "printf" on the data.  If the data happens to begin with a dash
character, "printf" would interpret it as a flag argument and might fail.
It is therefore more convenient in these cases to use the regular form
of Base64 encoding.

Conversely, if we only generated data encoded in the regular form of
Base64, we would not be able to use it in the several places where
we generate random directory names without also passing the data through
something like the sha256sum(1) program, as we do now.  By making
the URL-compatible encoding available directly from our new utility
program, we can simplify these portions of our test suite when, in a
subsequent commit in this PR, we refactor our test's shell functions
to make use of our new program.

Note that we follow the example of several of our other test helper
programs, such as lfstest-realpath and lfstest-testutils, in using
distinct exit codes (specifically, the values 2 and above) for each
error condition.