t: use lfstest-genrandom program in tests
In a previous commit in this PR we introduced a new utility program
for our test suite, lfstest-genrandom, which reliably produces random
data on all the platforms we support, either as raw binary data or
using one of several forms of Base64 encoding.

We now make use of that program throughout our test suite, in place
of reading from /dev/urandom directly.  (Note that because the
lfstest-genrandom program uses the Go crypto/rand package to collect
the random data, on macOS and Linux systems the ultimate source of the
data will still be /dev/urandom, since the Go package reads from
that special file on those platforms.  On Windows, however, the Go
package makes use of the native Windows ProcessPrng API, so we will
no longer be dependent on the Git Bash environment to provide a
simulated /dev/urandom on Windows.)

In most cases, we pass the --base64 flag to our lfstest-genrandom
program, which allows us to eliminate many calls to the base64(1)
command.  We can also pass the exact size of the returned data
we desire as a command-line argument, which eliminates the need
to pipe the data through the head(1) command.

In two instances, in the remove_and_create_local_repo() and
remove_and_create_remote_repo() functions in our t/fixtures/migrate.sh
file, we use the --base64url flag so that we can use the returned
strings as directory name suffixes without further processing.
Previously, we had to pipe the data through base64(1), head(1),
sha256sum(1), and cut(1) in order to ensure we had a random string
safe for use in a directory name.

And in the setup_local_branch_with_special_character_files() function
in our t/fixtures/migrate.sh file we don't pass any flags to the
lfstest-genrandom program as we want raw binary data for the test
files created by that function.