t/t-prune.sh: use lfstest-genrandom not uuidgen
In previous commits in this PR we introduced a new utility program
for our test suite, lfstest-genrandom, which reliably produces random
data on all the platforms we support (in several forms of Base64
encoding, or as raw binary data), and then converted much of our test
suite to make use of this new program.

As noted in one of those commits, one test in our t/t-prune.sh test
script makes use of the uuidgen(1) command to populate some Git LFS
objects with arbitrary text, but as this command does not exist in
the Git Bash environment on Windows, the objects actually have empty
data on that platform.  The test still passes because it doesn't depend
on the length of the random data, and because the non-zero exit code
caused by the missing uuidgen(1) command is ignored as it appears earlier
in a pipeline whose final command (our lfstest-testutils program)
always succeeds.

We therefore revise this "prune verify large numbers of refs" test
(which was added in commit dd19c1b9c2d281b986829424ff57bd2bf3be2f29 of
PR #1861) to call our new lfstest-genrandom program instead of the
uuidgen(1) command.  We pass the --base64 flag to our program and request
a slightly larger size for the output Base64-encoded string than the
length of a UUID, just to ensure we are using a similar amount of random
data as before, although this doesn't actually affect the success or
failure of the test.

We also adjust the test so that it does not create the Git LFS objects
with sizes that do not match the length of their arbitrary content.
(Again, the test passed despite this inconsistency in its test data.)