script/cibuild,t: always enable test count locks
In commit 9b73c8096ac546de0a1b076453ec548b1c5bc7ae of PR #3125 the
GIT_LFS_LOCK_ACQUIRE_DISABLED environment variable was introduced
for use with our test suite.  When this variable is set to the value
"1", it indicates to the lfstest-count-tests test utility that it
should skip creating a test_count.lock file before updating the
test_count file we use to track the number of active test scripts.

This was done because the lfstest-count-tests program had a bug which
prevented it from deleting the lock file on Windows, causing subsequent
invocations of the program to be unable to exclusively create it
again.  As this bug was not resolved at the time, the choice was
made to work around the problem by skipping the creation of a
lock file altogether on Windows.

This change was also made in conjunction with the addition of another
environment variable, GIT_LFS_NO_TEST_COUNT, in commit
c591ff7fbb738e897dae9f821b0d4b37aba70b00 of the same PR.  That variable
is also set only on Windows, and stops the setup() and shutdown()
shell functions defined in our t/testhelpers.sh script from calling
the lfstest-count-tests program.

As we have now resolved the underlying problem in the lfstest-count-tests
utility, in a prior commit in this PR, we can now remove both of
these environment variables and the test suite features they control.

We start by removing support for the GIT_LFS_LOCK_ACQUIRE_DISABLED
environment variable from the lfstest-count-tests utility, and
changing our script/cibuild script so it no longer sets the variable.

We can also remove the environment variable from the list of those
we unset before running certain tests.  We had to introduce this
behaviour in commit aca1aff48e0c75609de756a0e3640071442bc5e3 of
PR #3808, when we migrated our CI jobs to GitHub Actions.  However,
we will no longer need to unset these variables as we will now never
set them at all, and they will have no effect if they were set.