script/cibuild,t: always count running tests
In commit c591ff7fbb738e897dae9f821b0d4b37aba70b00 of PR #3125 the
GIT_LFS_NO_TEST_COUNT environment variable was introduced for use
with our test suite.  When this variable is set to a non-empty
value, it indicates to the setup() and shutdown() test helper
shell functions that they should skip running the lfstest-count-tests
utility program entirely.

This was done because the lfstest-count-tests program had a bug which
prevented it from deleting the lock file on Windows, causing subsequent
invocations of the program to be unable to exclusively create it
again.  As this bug was not resolved at the time, the choice was
made to work around the problem in part by not counting the number
of running test scripts on Windows.  (We also introduced the
GIT_LFS_LOCK_ACQUIRE_DISABLED environment variable, in commit
9b73c8096ac546de0a1b076453ec548b1c5bc7ae of the same PR, to further
address the problem.)

When running our test suite on Windows, we set the GIT_LFS_NO_TEST_COUNT
to "1", which causes the individual test scripts to skip invoking
lfstest-count-tests entirely.  Instead, the program is run just once
in the "test" t/Makefile target before executing the test scripts with
the "prove" command, and then once again afterwards, both times by
explicitly resetting the GIT_LFS_NO_TEST_COUNT variable to an empty
string value.  These two invocations cause the lfstest-count-tests
program to start the lfstest-gitserver program before all the test
scripts are executed and then stop it once they have all finished.

As we have now resolved the underlying problem in the lfstest-count-tests
utility, in a prior commit in this PR, we can remove this environment
variable and the test suite features it controls.

We change our script/cibuild script to no longer set the
GIT_LFS_NO_TEST_COUNT variable, and we revise the conditionals
in the setup() and shutdown() test helper shell functions so they
always call the lfstest-count-tests program, rather than only doing
so if the GIT_LFS_NO_TEST_COUNT variable is unset or empty.  We
also change the "test" target's recipe in our t/Makefile file to
not unset GIT_LFS_NO_TEST_COUNT when running lfstest-count-tests
before and after it runs the "prove" command.

Finally, as we have already removed the GIT_LFS_LOCK_ACQUIRE_DISABLED
environment variable and all support for it, in a previous commit in
this PR, we are now also able to eliminate the unset_vars() shell
functions in some of our t/t-*.sh test scripts.  These functions are used
to unset both the GIT_LFS_LOCK_ACQUIRE_DISABLED and GIT_LFS_NO_TEST_COUNT
variables before running the tests in those scripts.  We added this
behaviour in commit aca1aff48e0c75609de756a0e3640071442bc5e3 of PR #3808,
when we migrated our CI jobs to GitHub Actions.  However, we will no
longer need to unset these variables as we will now never set them at all,
and they will have no effect if they are set.